feat(vaults): check bitcoin cosign sig
feat(vaults): publish market rate of bitcoin
chore: refactor bitcoin into own package

feat(bond): verify bitcoin signatures
