+++ bitcoin/Cargo.toml
@@ -1,0 +1,59 @@
+[package]
+name = "ulx-bitcoin"
+description = "Modules for creating an unlocking timeline scripts for bitcoin vaults"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+homepage.workspace = true
+license.workspace = true
+rust-version.workspace = true
+repository.workspace = true
+documentation.workspace = true
+
+[dependencies]
+thiserror = { workspace = true, optional = true }
+codec = { workspace = true, default-features = false }
+
+sp-runtime = { workspace = true, default-features = false }
+sp-std = { workspace = true, default-features = false }
+sp-core = { workspace = true, default-features = false }
+sp-io = { workspace = true, default-features = false }
+bitcoin = { workspace = true, default-features = false }
+k256 = { workspace = true, features = ["alloc", "ecdsa"], default-features = false }
+bitcoincore-rpc = { workspace = true, optional = true }
+
+anyhow = { workspace = true, optional = true }
+
+parking_lot = { workspace = true, optional = true }
+hwi = { workspace = true, optional = true }
+
+ulx-primitives = { workspace = true, default-features = false, features = ["bitcoin"] }
+
+[dev-dependencies]
+frame-support = { workspace = true }
+sp-keyring = { workspace = true }
+ulx-testing = { workspace = true }
+bitcoind = { workspace = true }
+env_logger = { workspace = true }
+rand = { workspace = true }
+
+
+[features]
+default = [ "std" ]
+std = [
+	"anyhow/std",
+	"bitcoin/std",
+	"bitcoincore-rpc/default",
+	"codec/std",
+	"k256/std",
+	"parking_lot/default",
+	"sp-core/std",
+	"sp-io/std",
+	"sp-runtime/std",
+	"sp-std/std",
+	"thiserror",
+	"ulx-primitives/bitcoin",
+	"ulx-primitives/std",
+]
+fast-runtime = [ "ulx-primitives/fast-runtime" ]
+hwi = [ "dep:hwi" ]
