+++ bitcoin/cli/src/create_vault.rs
@@ -1,0 +1,3 @@
+pub fn vault_creation_cli() {
+    println!("vault_creation_cli");
+}
