+++ bitcoin/src/lib.rs
@@ -1,0 +1,14 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+
+pub use cosign_script::{Amount, CosignScript, UnlockStep};
+pub use errors::Error;
+#[cfg(feature = "std")]
+pub use utxo_spend_filter::{BlockFilter, UtxoSpendFilter};
+pub use utxo_unlocker::UtxoUnlocker;
+
+mod cosign_script;
+mod utxo_unlocker;
+
+mod errors;
+#[cfg(feature = "std")]
+mod utxo_spend_filter;
