+++ bitcoin/src/utxo_unlocker.rs
@@ -1,6 +1,3 @@
-use std::collections::BTreeMap;
-
-use anyhow::{anyhow, bail, ensure};
 use bitcoin::{
 	absolute::LockTime,
 	bip32::{KeySource, Xpriv},
@@ -8,112 +5,31 @@ use bitcoin::{
 	key::Secp256k1,
 	psbt::Input,
 	sighash::SighashCache,
-	transaction::{predict_weight, InputWeightPrediction, Version},
-	Address, Amount, EcdsaSighashType, FeeRate, Network, OutPoint, PrivateKey, Psbt, PubkeyHash,
-	PublicKey, ScriptBuf, Sequence, Transaction, TxIn, TxOut, Witness,
+	transaction::Version,
+	Amount, EcdsaSighashType, OutPoint, PrivateKey, Psbt, PubkeyHash, PublicKey, ScriptBuf,
+	Sequence, Transaction, TxIn, TxOut, Witness,
+};
+use k256::ecdsa::signature::Verifier;
+use sp_std::{vec, vec::Vec};
+
+use ulx_primitives::{
+	bitcoin::{
+		BitcoinError, BitcoinHeight, BitcoinPubkeyHash, BitcoinSignature, CompressedBitcoinPubkey,
+		Satoshis,
+	},
+	ensure,
 };
 
-use ulx_primitives::bitcoin::{
-	create_timelock_multisig_script, BitcoinHeight, BitcoinPubkeyHash, Satoshis,
+use crate::{
+	cosign_script::{CosignScript, UnlockStep, COSIGN_CODE, OPEN_CLAIM_CODE, VAULT_CLAIM_CODE},
+	errors::Error,
 };
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct CosignScript {
-	vault_pubkey_hash: BitcoinPubkeyHash,
-	owner_pubkey_hash: BitcoinPubkeyHash,
-	vault_claim_height: BitcoinHeight,
-	open_claim_height: BitcoinHeight,
-	created_at_height: BitcoinHeight,
-	script: ScriptBuf,
-}
-
-impl CosignScript {
-	pub fn new(
-		vault_pubkey_hash: BitcoinPubkeyHash,
-		owner_pubkey_hash: BitcoinPubkeyHash,
-		vault_claim_height: BitcoinHeight,
-		open_claim_height: BitcoinHeight,
-		created_at_height: BitcoinHeight,
-	) -> anyhow::Result<Self> {
-		let script = create_timelock_multisig_script(
-			vault_pubkey_hash,
-			owner_pubkey_hash,
-			vault_claim_height,
-			open_claim_height,
-		)
-		.map_err(|_| anyhow!("Unable to create the timelock script"))?;
-		Ok(Self {
-			vault_pubkey_hash,
-			owner_pubkey_hash,
-			vault_claim_height,
-			open_claim_height,
-			created_at_height,
-			script,
-		})
-	}
-
-	pub fn get_script_address(&self, network: Network) -> Address {
-		Address::p2wsh(&self.script, network)
-	}
-
-	pub fn get_script_pubkey(&self) -> ScriptBuf {
-		self.script.clone().to_p2wsh()
-	}
-
-	pub fn set_registered_height(&mut self, height: BitcoinHeight) {
-		self.created_at_height = height;
-	}
-
-	pub fn unlock_height(&self, unlock_step: UnlockStep) -> u32 {
-		(match unlock_step {
-			UnlockStep::OwnerCosign | UnlockStep::VaultCosign => self.created_at_height,
-			UnlockStep::VaultClaim => self.vault_claim_height,
-			UnlockStep::OwnerClaim => self.open_claim_height,
-		}) as u32
-	}
-	pub fn calculate_fee(
-		&self,
-		is_cosign: bool,
-		to_script_pubkey: ScriptBuf,
-		fee_rate: FeeRate,
-	) -> anyhow::Result<Amount> {
-		const MAX_SIGNATURE_SIZE: usize = 73;
-		const COMPRESSED_PUBKEY_SIZE: usize = 33;
-		const UNLOCK_CODE_SIZE: usize = 1;
-
-		let mut witness_element_lengths =
-			vec![MAX_SIGNATURE_SIZE, COMPRESSED_PUBKEY_SIZE, UNLOCK_CODE_SIZE, self.script.len()];
-
-		if is_cosign {
-			witness_element_lengths.push(MAX_SIGNATURE_SIZE);
-			witness_element_lengths.push(COMPRESSED_PUBKEY_SIZE);
-		}
-		let weight = predict_weight(
-			vec![InputWeightPrediction::from_slice(0, witness_element_lengths.as_slice())],
-			vec![to_script_pubkey.len()],
-		);
-		let Some(fee) = fee_rate.fee_wu(weight) else { bail!("Fee rate too low") };
-		Ok(fee)
-	}
-}
 
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct UtxoUnlocker {
 	pub cosign_script: CosignScript,
 	pub unlock_step: UnlockStep,
 	pub psbt: Psbt,
-}
-
-pub const COSIGN_CODE: u8 = 1;
-pub const VAULT_CLAIM_CODE: u8 = 2;
-pub const OPEN_CLAIM_CODE: u8 = 3;
-
-#[derive(Copy, Clone, Debug, Eq, Hash, PartialEq, PartialOrd, Ord)]
-pub enum UnlockStep {
-	VaultCosign,
-	OwnerCosign,
-	VaultClaim,
-	OwnerClaim,
 }
 
 impl UtxoUnlocker {
@@ -125,28 +41,26 @@ impl UtxoUnlocker {
 		unlock_step: UnlockStep,
 		fee: Amount,
 		to_script_pubkey: ScriptBuf,
-	) -> anyhow::Result<Self> {
+	) -> Result<Self, Error> {
 		let lock_time = cosign_script.unlock_height(unlock_step);
 		let out_point = OutPoint { txid: utxo_txid, vout: utxo_vout };
 		let out_amount = Amount::from_sat(utxo_satoshis);
 		let unsigned_tx = Transaction {
 			version: Version::TWO, // Post BIP-68.
 			lock_time: LockTime::from_height(lock_time)
-				.map_err(|_| anyhow!("Invalid lock time"))?,
+				.map_err(|_| BitcoinError::InvalidLockTime)?,
 			input: vec![TxIn {
 				previous_output: out_point,
 				sequence: Sequence::ENABLE_LOCKTIME_NO_RBF,
 				..TxIn::default()
 			}],
 			output: vec![TxOut {
-				value: out_amount
-					.checked_sub(fee)
-					.ok_or(anyhow!("Fee caused overflow of amount"))?,
+				value: out_amount.checked_sub(fee).ok_or(Error::FeeOverflow)?,
 				script_pubkey: to_script_pubkey,
 			}],
 		};
 
-		let mut psbt = Psbt::from_unsigned_tx(unsigned_tx)?;
+		let mut psbt = Psbt::from_unsigned_tx(unsigned_tx).map_err(Error::PsbtError)?;
 
 		psbt.inputs[0] = Input {
 			witness_utxo: Some(TxOut {
@@ -174,7 +88,7 @@ impl UtxoUnlocker {
 		unlock_step: UnlockStep,
 		fee: Amount,
 		pay_to_script_pubkey: ScriptBuf,
-	) -> anyhow::Result<Self> {
+	) -> Result<Self, Error> {
 		Self::from_script(
 			CosignScript::new(
 				vault_pubkey_hash.into(),
@@ -196,10 +110,37 @@ impl UtxoUnlocker {
 		self.psbt.inputs[0].partial_sigs.insert(pubkey, signature);
 	}
 
-	pub fn sign(&mut self, privkey: PrivateKey) -> anyhow::Result<(Signature, PublicKey)> {
+	/// No std friendly version of verifying a signature
+	pub fn verify_signature_raw(
+		&self,
+		pubkey: CompressedBitcoinPubkey,
+		signature_der_bytes: &BitcoinSignature,
+	) -> Result<bool, Error> {
+		let psbt = &self.psbt;
+		let mut cache = SighashCache::new(&psbt.unsigned_tx);
+
+		// Get the sighash message
+		let (msg, _) = match psbt.sighash_ecdsa(0, &mut cache) {
+			Ok(result) => result,
+			Err(_) => return Ok(false),
+		};
+
+		let (_sighash_type, sigdata) =
+			signature_der_bytes.0.split_last().ok_or(Error::InvalidSignatureBytes)?;
+
+		let signature =
+			k256::ecdsa::Signature::from_der(sigdata).map_err(|_| Error::InvalidSignatureBytes)?;
+
+		let pubkey = k256::ecdsa::VerifyingKey::from_sec1_bytes(&pubkey.0)
+			.map_err(|_| Error::InvalidCompressPubkeyBytes)?;
+
+		Ok(pubkey.verify(msg.as_ref(), &signature).is_ok())
+	}
+
+	pub fn sign(&mut self, privkey: PrivateKey) -> Result<(Signature, PublicKey), Error> {
 		let psbt = &mut self.psbt;
 		let mut cache = SighashCache::new(&psbt.unsigned_tx);
-		let (msg, ecdsa_type) = psbt.sighash_ecdsa(0, &mut cache)?;
+		let (msg, ecdsa_type) = psbt.sighash_ecdsa(0, &mut cache).map_err(Error::SignError)?;
 		let secp = Secp256k1::new();
 		let sig = secp.sign_ecdsa(&msg, &privkey.inner);
 		let signature = Signature { signature: sig, sighash_type: ecdsa_type };
@@ -212,30 +153,32 @@ impl UtxoUnlocker {
 		&mut self,
 		master_xpriv: Xpriv,
 		key_source: KeySource,
-	) -> anyhow::Result<(Signature, PublicKey)> {
+	) -> Result<(Signature, PublicKey), Error> {
 		let psbt = &mut self.psbt;
 		let secp = Secp256k1::new();
-		let child_xpriv = master_xpriv.derive_priv(&secp, &key_source.1)?;
+		let child_xpriv = master_xpriv
+			.derive_priv(&secp, &key_source.1)
+			.map_err(|e| Error::Bip32Error(e))?;
 		let child_priv = child_xpriv.to_priv();
 		let pubkey = child_priv.public_key(&secp);
 
 		psbt.inputs[0].bip32_derivation.insert(pubkey.inner, key_source);
 
 		match psbt.sign(&master_xpriv, &secp) {
-			Ok(_) => ensure!(!psbt.inputs[0].partial_sigs.is_empty(), "Expected 1 signature"),
-			Err((_, errs)) => bail!("Error signing Partially Signed Bitcoin script {:?}", errs),
+			Ok(_) => ensure!(!psbt.inputs[0].partial_sigs.is_empty(), Error::SignatureExpected),
+			Err((_, errs)) => return Err(Error::SigningErrors(errs)),
 		};
 
 		let Some((_, signature)) =
 			psbt.inputs[0].partial_sigs.iter().find(|(k, _)| k.inner == pubkey.inner)
 		else {
-			bail!("Could not sign with derived key");
+			return Err(Error::DerivedKeySignError);
 		};
 
 		Ok((*signature, pubkey))
 	}
 
-	pub fn extract_tx(&mut self) -> anyhow::Result<Transaction> {
+	pub fn extract_tx(&mut self) -> Result<Transaction, Error> {
 		let mut witness = Witness::new();
 		let psbt = &mut self.psbt;
 		let mut sigs: Vec<Vec<u8>> = vec![];
@@ -249,7 +192,7 @@ impl UtxoUnlocker {
 				sigs.insert(0, sig.to_vec());
 				sigs.insert(1, pubkey.to_bytes());
 			} else {
-				bail!("Unknown pubkey hash in partial sigs");
+				return Err(Error::UnknownPubkeyHash);
 			}
 		}
 		for sig in sigs {
@@ -265,26 +208,28 @@ impl UtxoUnlocker {
 		psbt.inputs[0].final_script_witness = Some(witness);
 
 		// Clear all the data fields as per the spec.
-		psbt.inputs[0].partial_sigs = BTreeMap::new();
+		psbt.inputs[0].partial_sigs.clear();
 		psbt.inputs[0].sighash_type = None;
 		psbt.inputs[0].redeem_script = None;
 		psbt.inputs[0].witness_script = None;
-		psbt.inputs[0].bip32_derivation = BTreeMap::new();
+		psbt.inputs[0].bip32_derivation.clear();
 
-		let tx = psbt.clone().extract_tx()?;
+		let tx = psbt.clone().extract_tx().map_err(|e| Error::ExtractTxError(e))?;
 		Ok(tx)
 	}
 }
+
 #[cfg(feature = "hwif")]
 mod hwi {
 	use hwi::{types::HWIDevice, HWIClient};
+
 	impl UtxoUnlocker {
 		pub fn sign_hwi(
 			&mut self,
 			key_source: KeySource,
 			device: Option<HWIDevice>,
 			network: Network,
-		) -> anyhow::Result<(Signature, PublicKey)> {
+		) -> Result<(Signature, PublicKey)> {
 			let psbt = &mut self.psbt;
 			let mut device = device;
 			if device.is_none() {
