+++ end-to-end/Cargo.toml
@@ -20,6 +20,7 @@ bitcoin = { workspace = true }
 lazy_static = { workspace = true }
 ulx-primitives = { workspace = true, features = ["default"] }
 ulx-bitcoin-utxo-tracker = { workspace = true }
+ulx-bitcoin = { workspace = true }
 ulixee-client = { workspace = true }
 sp-arithmetic = { workspace = true }
 sp-core = { workspace = true }
