+++ node/bitcoin_utxo_tracker/Cargo.toml
@@ -23,11 +23,11 @@ thiserror = { workspace = true }
 bitcoincore-rpc = { workspace = true }
 bitcoin = { workspace = true }
 ulx-primitives = { workspace = true, features = ["std", "bitcoin"] }
+ulx-bitcoin = { workspace = true, features = ["std"] }
 sc-service = { workspace = true }
 sp-api = { workspace = true, features = ["std"] }
 sp-runtime = { workspace = true, features = ["std"] }
 sc-client-api = { workspace = true }
-hwi = { workspace = true, optional = true }
 
 [dev-dependencies]
 bitcoind = { workspace = true }
@@ -36,7 +36,3 @@ tokio = { workspace = true }
 rand = { workspace = true }
 lazy_static = { workspace = true }
 ulx-testing = { workspace = true }
-
-[features]
-default = []
-hwi = [ "dep:hwi" ]
