+++ oracle/src/btc_price.rs
@@ -34,7 +34,7 @@ impl BtcPriceLookup {
 			let mut mock = MOCK_BTC_PRICE.lock().unwrap();
 			if let Some(x) = mock.take() {
 				*mock = Some(x.clone());
-				return Ok(x.clone())
+				return Ok(x.clone());
 			}
 		}
