+++ oracle/src/main.rs
@@ -8,7 +8,6 @@ use sp_core::{
 	sr25519, Pair as PairT,
 };
 use sp_runtime::traits::IdentifyAccount;
-use tracing::warn;
 use ulixee_client::signer::KeystoreSigner;
 use ulx_primitives::{AccountId, CryptoType, KeystoreParams, ADDRESS_PREFIX};
 use url::Url;
@@ -26,13 +25,17 @@ pub(crate) mod utils;
 #[derive(Parser, Debug)]
 #[clap(version = crate_version!())]
 #[command(author, version, about, arg_required_else_help = true, long_about = None)]
+#[clap(arg_required_else_help = true)]
 struct Cli {
+	#[command(subcommand)]
+	pub subcommand: Subcommand,
+
 	/// Start in dev mode (using default //Alice as operator)
-	#[clap(long)]
+	#[clap(global = true, long)]
 	dev: bool,
 
 	/// What mainchain RPC websocket url do you want to reach out use to sync blocks?
-	#[clap(short, long, env, default_value = "ws://127.0.0.1:9944")]
+	#[clap(global = true, short, long, env, default_value = "ws://127.0.0.1:9944")]
 	trusted_rpc_url: String,
 
 	#[allow(missing_docs)]
@@ -40,15 +43,12 @@ struct Cli {
 	keystore_params: KeystoreParams,
 
 	/// The signer to use from the keystore (Required if not in dev mode)
-	#[clap(long, env)]
+	#[clap(global = true, long, env)]
 	signer_address: Option<String>,
 
 	/// What type of crypto to use for the signer (Required if not in dev mode)
-	#[clap(long, env)]
+	#[clap(global = true, long, env)]
 	signer_crypto: Option<OracleCryptoType>,
-
-	#[command(subcommand)]
-	pub subcommand: Option<Subcommand>,
 }
 #[derive(Debug, Clone, clap::Subcommand, Default)]
 #[allow(clippy::large_enum_variant)]
@@ -80,7 +80,10 @@ impl From<OracleCryptoType> for CryptoType {
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
 	let _ = tracing_subscriber::FmtSubscriber::builder()
-		.with_env_filter(tracing_subscriber::EnvFilter::from("debug"))
+		.with_env_filter(
+			tracing_subscriber::EnvFilter::try_from_default_env()
+				.unwrap_or(tracing_subscriber::EnvFilter::from("info")),
+		)
 		.try_init();
 	env::set_var("RUST_BACKTRACE", "1");
 	dotenv().ok();
@@ -94,9 +97,6 @@ async fn main() -> anyhow::Result<()> {
 		trusted_rpc_url,
 		if dev { "Dev Mode" } else { "" }
 	);
-	if subcommand.is_none() {
-		warn!("No subcommand provided, defaulting to price index. NOTE: the bitcoin oracle must be activated somewhere else.");
-	}
 
 	let keystore = if dev {
 		keystore_params.open_dev("//Alice", CryptoType::Sr25519, ACCOUNT)?
@@ -124,9 +124,8 @@ async fn main() -> anyhow::Result<()> {
 
 	let signer = KeystoreSigner::new(keystore, signer_account, signer_crypto.into());
 	match subcommand {
-		Some(Subcommand::PriceIndex) | None =>
-			price_index_loop(trusted_rpc_url, signer, dev).await?,
-		Some(Subcommand::Bitcoin { bitcoin_rpc_url }) => {
+		Subcommand::PriceIndex => price_index_loop(trusted_rpc_url, signer, dev).await?,
+		Subcommand::Bitcoin { bitcoin_rpc_url } => {
 			let bitcoin_url = Url::parse(&bitcoin_rpc_url).map_err(|e| {
 				anyhow!("Unable to parse bitcoin rpc url ({}) {:?}", bitcoin_rpc_url, e)
 			})?;
