+++ pallets/bond/Cargo.toml
@@ -26,6 +26,7 @@ sp-runtime = { workspace = true }
 sp-core = { workspace = true }
 sp-std = { workspace = true }
 ulx-primitives = { workspace = true, features = ["bitcoin"] }
+ulx-bitcoin = { workspace = true, default-features = false }
 
 [dev-dependencies]
 pallet-balances = { workspace = true }
@@ -47,6 +48,8 @@ std = [
 	"sp-io/std",
 	"sp-runtime/std",
 	"sp-std/std",
+	"ulx-bitcoin/std",
+	"ulx-primitives/bitcoin",
 	"ulx-primitives/std",
 ]
 runtime-benchmarks = [
