+++ pallets/vaults/Cargo.toml
@@ -25,6 +25,7 @@ sp-runtime = { workspace = true }
 sp-core = { workspace = true }
 sp-std = { workspace = true }
 ulx-primitives = { workspace = true, features = ["bitcoin"] }
+ulx-bitcoin = { workspace = true, default-features = false }
 
 [dev-dependencies]
 pallet-balances = { workspace = true }
@@ -45,6 +46,7 @@ std = [
 	"sp-io/std",
 	"sp-runtime/std",
 	"sp-std/std",
+	"ulx-bitcoin/std",
 	"ulx-primitives/std",
 ]
 runtime-benchmarks = [
