+++ primitives/src/apis.rs
@@ -74,6 +74,7 @@ sp_api::decl_runtime_apis! {
 		fn get_sync_status() -> Option<BitcoinSyncStatus>;
 		fn active_utxos() -> Vec<(Option<UtxoRef>, UtxoValue)>;
 		fn redemption_rate(satoshis: Satoshis) -> Option<Balance>;
+		fn market_rate(satoshis: Satoshis) -> Option<Balance>;
 	}
 }
