+++ primitives/src/bitcoin.rs
@@ -70,77 +70,6 @@ impl TryFrom<Vec<u8>> for BitcoinSignature {
 	}
 }
 
-/// Creates a bitcoin script the does the following:
-/// - Until `vault_claim_height`, multisig requires both public keys and signatures to be revealed
-/// - Between `vault_claim_height` and `open_claim_height`, only the vault can claim the funds
-/// - After `open_claim_height`, either party can claim the funds
-#[cfg(feature = "bitcoin")]
-#[rustfmt::skip]
-pub fn create_timelock_multisig_script(
-	vault_pubkey_hash: BitcoinPubkeyHash,
-	owner_pubkey_hash: BitcoinPubkeyHash,
-	vault_claim_height: BitcoinHeight,
-	open_claim_height: BitcoinHeight,
-) -> Result<bitcoin::ScriptBuf, BitcoinError> {
-	use bitcoin::blockdata::{opcodes::all::*, script::Builder};
-	use bitcoin::absolute::LockTime;
-
-	let script = Builder::new()
-		// code 1 is unlock
-		.push_opcode(OP_DUP)
-		.push_int(1)
-		.push_opcode(OP_EQUAL)
-		.push_opcode(OP_IF)
-			.push_opcode(OP_DROP)
-			.push_opcode(OP_DUP)
-			.push_opcode(OP_HASH160)
-			.push_slice(vault_pubkey_hash.0)
-			// set 1 to stack if this is the vault
-			.push_opcode(OP_EQUALVERIFY)
-			.push_opcode(OP_CHECKSIGVERIFY)
-
-			// now consume user key
-			.push_opcode(OP_DUP)
-			.push_opcode(OP_HASH160)
-			.push_slice(owner_pubkey_hash.0)
-			//  OP_EQUALVERIFY OP_CHECKSIG at end
-
-		.push_opcode(OP_ELSE)
-			.push_int(2)
-			.push_opcode(OP_EQUAL)
-		    // code 2 is vault claim
-			.push_opcode(OP_IF)
-				.push_lock_time(LockTime::from_height(vault_claim_height as u32).map_err(|_| BitcoinError::InvalidLockTime)?)
-				.push_opcode(OP_CLTV)
-				.push_opcode(OP_DROP)
-
-				.push_opcode(OP_DUP)
-				.push_opcode(OP_HASH160)
-				.push_slice(vault_pubkey_hash.0)
-				//  OP_EQUALVERIFY OP_CHECKSIG at end
-
-			// code 3 is owner claim
-			.push_opcode(OP_ELSE)
-				.push_lock_time(LockTime::from_height(open_claim_height as u32).map_err(|_| BitcoinError::InvalidLockTime)?)
-				.push_opcode(OP_CLTV)
-				.push_opcode(OP_DROP)
-
-				.push_opcode(OP_DUP)
-				.push_opcode(OP_HASH160)
-				.push_slice(owner_pubkey_hash.0)
-				//  OP_EQUALVERIFY OP_CHECKSIG at end
-			.push_opcode(OP_ENDIF)
-		.push_opcode(OP_ENDIF)
-
-
-		.push_opcode(OP_EQUALVERIFY)
-		.push_opcode(OP_CHECKSIG)
-
-
-		.into_script();
-	Ok(script)
-}
-
 /// A Script Pubkey for a Bitcoin UTXO. Supported types are:
 /// - P2WSH (Pay to Witness Script Hash)
 #[derive(Clone, PartialEq, Eq, Encode, Decode, TypeInfo, RuntimeDebug, Copy)]
