+++ primitives/src/providers.rs
@@ -10,7 +10,7 @@ use sp_std::vec::Vec;
 
 use crate::{
 	bitcoin::{
-		BitcoinCosignScriptPubkey, BitcoinHeight, BitcoinRejectedReason, Satoshis, UtxoId,
+		BitcoinCosignScriptPubkey, BitcoinHeight, BitcoinRejectedReason, Satoshis, UtxoId, UtxoRef,
 		SATOSHIS_PER_BITCOIN,
 	},
 	block_seal::{BlockPayout, MiningAuthority, RewardSharing},
@@ -69,8 +69,7 @@ pub trait PriceProvider<Balance: Codec + AtLeast32BitUnsigned> {
 	fn get_latest_argon_price_in_us_cents() -> Option<FixedU128>;
 
 	/// The argon CPI is the US CPI deconstructed by the Argon market price in Dollars.
-	/// This value has 3 decimal places of precision in a whole number (eg, 1 = 1_000, -1 = -1_000)
-	fn get_argon_cpi_price() -> Option<ArgonCPI>;
+	fn get_argon_cpi() -> Option<ArgonCPI>;
 }
 
 pub trait BitcoinUtxoTracker {
@@ -81,6 +80,7 @@ pub trait BitcoinUtxoTracker {
 		satoshis: Satoshis,
 		watch_for_spent_until: BitcoinHeight,
 	) -> Result<(), DispatchError>;
+	fn get(utxo_id: UtxoId) -> Option<UtxoRef>;
 	fn unwatch(utxo_id: UtxoId);
 }
