+++ primitives/src/tick.rs
@@ -83,7 +83,7 @@ impl Ticker {
 		let now = now();
 		let current_tick = self.current();
 		let next_tick_time = self.time_for_tick(current_tick + 1);
-		let duration_to_next_tick = next_tick_time - now;
+		let duration_to_next_tick = next_tick_time.saturating_sub(now);
 		Duration::from_millis(duration_to_next_tick)
 	}
