+++ testing/src/bitcoind.rs
@@ -1,9 +1,18 @@
-use std::{env, fs::File};
+use std::{env, fs::File, str::FromStr};
 
 use anyhow::anyhow;
 use bitcoind::{downloaded_exe_path, BitcoinD, Conf};
 use fs2::FileExt;
 use lazy_static::lazy_static;
+use rand::{rngs::OsRng, RngCore};
+
+use bitcoin::{
+	bip32::{DerivationPath, Fingerprint, Xpriv, Xpub},
+	secp256k1::Secp256k1,
+	Address, Amount, CompressedPublicKey, Network, Txid,
+};
+use bitcoincore_rpc::{json::GetRawTransactionResult, RpcApi};
+use ulx_primitives::bitcoin::Satoshis;
 
 lazy_static! {
 	static ref BITCOIND_LOCK: std::sync::Mutex<()> = std::sync::Mutex::new(());
@@ -47,4 +56,84 @@ pub fn read_rpc_url(bitcoind: &BitcoinD) -> anyhow::Result<url::Url> {
 		url.set_password(Some(&cookie.password)).expect("password");
 	}
 	Ok(url)
+}
+
+pub fn wait_for_txid(
+	bitcoind: &BitcoinD,
+	txid: &Txid,
+	block_address: &Address,
+) -> GetRawTransactionResult {
+	loop {
+		// Attempt to get the raw transaction with verbose output
+		let result = bitcoind.client.call::<GetRawTransactionResult>(
+			"getrawtransaction",
+			&[txid.to_string().into(), 1.into()],
+		);
+
+		if let Ok(tx) = result {
+			if tx.confirmations.unwrap_or_default() > 1 {
+				return tx;
+			}
+		}
+		std::thread::sleep(std::time::Duration::from_secs(1));
+		add_blocks(bitcoind, 1, block_address);
+	}
+}
+
+pub fn create_xpriv() -> (Xpriv, Fingerprint) {
+	let secp = Secp256k1::new();
+	let mut seed = [0u8; 32];
+	OsRng.fill_bytes(&mut seed);
+	let master_xpriv = Xpriv::new_master(Network::Regtest, &seed).unwrap();
+	let master_xpub = Xpub::from_priv(&secp, &master_xpriv);
+	let fingerprint = master_xpub.fingerprint();
+	(master_xpriv, fingerprint)
+}
+
+pub fn derive(master_xpriv: &Xpriv, path: &str) -> (CompressedPublicKey, DerivationPath) {
+	let secp = Secp256k1::new();
+	let path = DerivationPath::from_str(path).expect("Invalid derivation path");
+	let child_xpriv = master_xpriv.derive_priv(&secp, &path).expect("Unable to derive child key");
+
+	let vault_privkey = child_xpriv.to_priv();
+	let vault_compressed_pubkey = CompressedPublicKey::from_private_key(&secp, &vault_privkey)
+		.expect("Unable to derive pubkey");
+	(vault_compressed_pubkey, path)
+}
+
+pub fn add_wallet_address(bitcoind: &BitcoinD) -> Address {
+	let address = bitcoind.client.get_new_address(None, None).unwrap();
+	address.require_network(Network::Regtest).unwrap()
+}
+
+pub fn add_blocks(bitcoind: &BitcoinD, count: u64, grant_to_address: &Address) {
+	bitcoind.client.generate_to_address(count, grant_to_address).unwrap();
+}
+
+pub fn fund_script_address(
+	bitcoind: &BitcoinD,
+	script_address: &Address,
+	amount: Satoshis,
+	block_address: &Address,
+) -> (Txid, u32, GetRawTransactionResult) {
+	let txid = bitcoind
+		.client
+		.send_to_address(
+			script_address,
+			Amount::from_sat(amount),
+			None,
+			None,
+			None,
+			None,
+			None,
+			None,
+		)
+		.unwrap();
+	let tx = wait_for_txid(bitcoind, &txid, block_address);
+	let vout = tx
+		.vout
+		.iter()
+		.position(|o| o.script_pub_key.script().unwrap() == script_address.script_pubkey())
+		.unwrap() as u32;
+	(txid, vout, tx)
 }
